+++ App/src/commands/context-menu/bot-check.js
@@ -11,14 +11,14 @@ module.exports = {
                 .setAuthor({ name: `${interaction.targetUser.tag} is a BOT!`, iconURL: `${interaction.targetUser.displayAvatarURL({ dynamic: true, size: 512 })}` })
                 .setColor(0xFFFFFF);
             message = await interaction.reply({ embeds: [embed], fetchReply: true });
-            message.react('🤖');
+            if (message.channel) message.react('🤖');
         }
         else {
             embed = new EmbedBuilder()
                 .setAuthor({ name: `${interaction.targetUser.tag} is not a BOT.`, iconURL: `${interaction.targetUser.displayAvatarURL({ dynamic: true, size: 512 })}` })
                 .setColor(0xFFFFFF);
             message = await interaction.reply({ embeds: [embed], fetchReply: true });
-            message.react('👤');
+            if (message.channel) message.react('👤');
         }
     },
 };