+++ App/src/commands/context-menu/kick.js
@@ -6,6 +6,6 @@ module.exports = {
         .setType(ApplicationCommandType.User)
         .setDefaultMemberPermissions(0),
     async execute(interaction) {
-        return interaction.reply({ content: `You wanted to kick: ${interaction.targetUser.username}`, ephemeral: true });
+        return interaction.reply({ content: `You wanted to kick: ${interaction.targetUser} :scales:`, ephemeral: true });
     },
 };