+++ App/src/commands/moderation/kick.js
@@ -5,13 +5,14 @@ module.exports = {
         .setName('kick')
         .setDescription('Select a member and kick them (but not really).')
         .setDefaultMemberPermissions(0)
+        .setDMPermission(false)
         .addUserOption(option => option
             .setName('target')
             .setDescription('The member to kick')
             .setRequired(true),
         ),
     async execute(interaction) {
-        const member = interaction.options.getMember('target');
-        return interaction.reply({ content: `You wanted to kick: ${member.user.username}`, ephemeral: true });
+        const user = interaction.options.getUser('target');
+        return interaction.reply({ content: `You wanted to kick: ${user} :scales:`, ephemeral: true });
     },
 };